refactor(identity): split UserService into focused partial files
Split the 682-line UserService.cs into smaller, focused partial class files:

- UserService.cs (~80 lines) - Core class with constructor and shared helpers
- UserService.Registration.cs (~220 lines) - RegisterAsync, GetOrCreateFromPrincipalAsync
- UserService.Profile.cs (~100 lines) - GetAsync, GetListAsync, UpdateAsync, existence checks
- UserService.Lifecycle.cs (~115 lines) - DeleteAsync, ToggleStatusAsync
- UserService.Roles.cs (~85 lines) - AssignRolesAsync, GetUserRolesAsync
- UserService.Confirmation.cs (~50 lines) - ConfirmEmailAsync, ConfirmPhoneNumberAsync

Existing partial files kept unchanged:
- UserService.Password.cs (87 lines)
- UserService.Permissions.cs (53 lines)

No interface or method signature changes. Build verified with 0 errors, 0 warnings.
