+++ src/Modules/Identity/Modules.Identity/Services/UserService.Profile.cs
@@ -1,0 +1,108 @@
+using FSH.Framework.Core.Exceptions;
+using FSH.Framework.Shared.Storage;
+using FSH.Framework.Storage;
+using FSH.Modules.Identity.Contracts.DTOs;
+using FSH.Modules.Identity.Domain;
+using Microsoft.EntityFrameworkCore;
+
+namespace FSH.Modules.Identity.Services;
+
+internal sealed partial class UserService
+{
+    public async Task<UserDto> GetAsync(string userId, CancellationToken cancellationToken)
+    {
+        var user = await userManager.Users
+            .AsNoTracking()
+            .Where(u => u.Id == userId)
+            .FirstOrDefaultAsync(cancellationToken);
+
+        _ = user ?? throw new NotFoundException("user not found");
+
+        return new UserDto
+        {
+            Id = user.Id,
+            Email = user.Email,
+            UserName = user.UserName,
+            FirstName = user.FirstName,
+            LastName = user.LastName,
+            ImageUrl = ResolveImageUrl(user.ImageUrl),
+            IsActive = user.IsActive
+        };
+    }
+
+    public Task<int> GetCountAsync(CancellationToken cancellationToken) =>
+        userManager.Users.AsNoTracking().CountAsync(cancellationToken);
+
+    public async Task<List<UserDto>> GetListAsync(CancellationToken cancellationToken)
+    {
+        var users = await userManager.Users.AsNoTracking().ToListAsync(cancellationToken);
+        var result = new List<UserDto>(users.Count);
+        foreach (var user in users)
+        {
+            result.Add(new UserDto
+            {
+                Id = user.Id,
+                Email = user.Email,
+                UserName = user.UserName,
+                FirstName = user.FirstName,
+                LastName = user.LastName,
+                ImageUrl = ResolveImageUrl(user.ImageUrl),
+                IsActive = user.IsActive
+            });
+        }
+
+        return result;
+    }
+
+    public async Task UpdateAsync(string userId, string firstName, string lastName, string phoneNumber, FileUploadRequest image, bool deleteCurrentImage)
+    {
+        var user = await userManager.FindByIdAsync(userId);
+
+        _ = user ?? throw new NotFoundException("user not found");
+
+        Uri imageUri = user.ImageUrl ?? null!;
+        if (image.Data != null || deleteCurrentImage)
+        {
+            var imageString = await storageService.UploadAsync<FshUser>(image, FileType.Image);
+            user.ImageUrl = new Uri(imageString, UriKind.RelativeOrAbsolute);
+            if (deleteCurrentImage && imageUri != null)
+            {
+                await storageService.RemoveAsync(imageUri.ToString());
+            }
+        }
+
+        user.FirstName = firstName;
+        user.LastName = lastName;
+        string? currentPhoneNumber = await userManager.GetPhoneNumberAsync(user);
+        if (phoneNumber != currentPhoneNumber)
+        {
+            await userManager.SetPhoneNumberAsync(user, phoneNumber);
+        }
+
+        var result = await userManager.UpdateAsync(user);
+        await signInManager.RefreshSignInAsync(user);
+
+        if (!result.Succeeded)
+        {
+            throw new CustomException("Update profile failed");
+        }
+    }
+
+    public async Task<bool> ExistsWithEmailAsync(string email, string? exceptId = null)
+    {
+        EnsureValidTenant();
+        return await userManager.FindByEmailAsync(email.Normalize()) is FshUser user && user.Id != exceptId;
+    }
+
+    public async Task<bool> ExistsWithNameAsync(string name)
+    {
+        EnsureValidTenant();
+        return await userManager.FindByNameAsync(name) is not null;
+    }
+
+    public async Task<bool> ExistsWithPhoneNumberAsync(string phoneNumber, string? exceptId = null)
+    {
+        EnsureValidTenant();
+        return await userManager.Users.FirstOrDefaultAsync(x => x.PhoneNumber == phoneNumber) is FshUser user && user.Id != exceptId;
+    }
+}
