+++ src/Modules/Identity/Modules.Identity/Services/UserService.Roles.cs
@@ -1,0 +1,94 @@
+using FSH.Framework.Core.Exceptions;
+using FSH.Framework.Shared.Constants;
+using FSH.Framework.Shared.Multitenancy;
+using FSH.Modules.Identity.Contracts.DTOs;
+using Microsoft.EntityFrameworkCore;
+
+namespace FSH.Modules.Identity.Services;
+
+internal sealed partial class UserService
+{
+    public async Task<string> AssignRolesAsync(string userId, List<UserRoleDto> userRoles, CancellationToken cancellationToken)
+    {
+        var user = await userManager.Users.Where(u => u.Id == userId).FirstOrDefaultAsync(cancellationToken);
+
+        _ = user ?? throw new NotFoundException("user not found");
+
+        // Check if the user is an admin for which the admin role is getting disabled
+        if (await userManager.IsInRoleAsync(user, RoleConstants.Admin)
+            && userRoles.Exists(a => !a.Enabled && a.RoleName == RoleConstants.Admin))
+        {
+            // Get count of users in Admin Role
+            int adminCount = (await userManager.GetUsersInRoleAsync(RoleConstants.Admin)).Count;
+
+            // Check if user is not Root Tenant Admin
+            // Edge Case : there are chances for other tenants to have users with the same email as that of Root Tenant Admin. Probably can add a check while User Registration
+            if (user.Email == MultitenancyConstants.Root.EmailAddress)
+            {
+                if (multiTenantContextAccessor?.MultiTenantContext?.TenantInfo?.Id == MultitenancyConstants.Root.Id)
+                {
+                    throw new CustomException("action not permitted");
+                }
+            }
+            else if (adminCount <= 2)
+            {
+                throw new CustomException("tenant should have at least 2 admins.");
+            }
+        }
+
+        var assignedRoles = new List<string>();
+
+        foreach (var userRole in userRoles)
+        {
+            // Check if Role Exists
+            if (await roleManager.FindByNameAsync(userRole.RoleName!) is not null)
+            {
+                if (userRole.Enabled)
+                {
+                    if (!await userManager.IsInRoleAsync(user, userRole.RoleName!))
+                    {
+                        await userManager.AddToRoleAsync(user, userRole.RoleName!);
+                        assignedRoles.Add(userRole.RoleName!);
+                    }
+                }
+                else
+                {
+                    await userManager.RemoveFromRoleAsync(user, userRole.RoleName!);
+                }
+            }
+        }
+
+        // Raise domain event for newly assigned roles
+        if (assignedRoles.Count > 0)
+        {
+            var tenantId = multiTenantContextAccessor?.MultiTenantContext?.TenantInfo?.Id;
+            user.RecordRolesAssigned(assignedRoles, tenantId);
+            await db.SaveChangesAsync(cancellationToken);
+        }
+
+        return "User Roles Updated Successfully.";
+
+    }
+
+    public async Task<List<UserRoleDto>> GetUserRolesAsync(string userId, CancellationToken cancellationToken)
+    {
+        var userRoles = new List<UserRoleDto>();
+
+        var user = await userManager.FindByIdAsync(userId);
+        if (user is null) throw new NotFoundException("user not found");
+        var roles = await roleManager.Roles.AsNoTracking().ToListAsync(cancellationToken);
+        if (roles is null) throw new NotFoundException("roles not found");
+        foreach (var role in roles)
+        {
+            userRoles.Add(new UserRoleDto
+            {
+                RoleId = role.Id,
+                RoleName = role.Name,
+                Description = role.Description,
+                Enabled = await userManager.IsInRoleAsync(user, role.Name!)
+            });
+        }
+
+        return userRoles;
+    }
+}
