+++ src/Modules/Identity/Modules.Identity/Services/UserService.cs
@@ -1,34 +1,19 @@
-﻿using Finbuckle.MultiTenant.Abstractions;
+using Finbuckle.MultiTenant.Abstractions;
 using FSH.Framework.Caching;
-using FSH.Framework.Core.Common;
-using FSH.Framework.Core.Exceptions;
 using FSH.Framework.Core.Context;
 using FSH.Framework.Eventing.Outbox;
 using FSH.Framework.Jobs.Services;
-using FSH.Framework.Mailing;
 using FSH.Framework.Mailing.Services;
-using FSH.Framework.Shared.Constants;
 using FSH.Framework.Shared.Multitenancy;
-using FSH.Framework.Storage;
-using FSH.Framework.Shared.Storage;
 using FSH.Framework.Storage.Services;
 using FSH.Framework.Web.Origin;
-using FSH.Modules.Identity.Contracts.DTOs;
-using FSH.Modules.Identity.Contracts.Events;
+using FSH.Modules.Auditing.Contracts;
 using FSH.Modules.Identity.Contracts.Services;
 using FSH.Modules.Identity.Data;
 using FSH.Modules.Identity.Domain;
-using FSH.Modules.Identity.Services;
-using FSH.Modules.Auditing.Contracts;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Options;
-using Microsoft.AspNetCore.WebUtilities;
-using Microsoft.EntityFrameworkCore;
-using System.Collections.ObjectModel;
-using System.Globalization;
-using System.Security.Claims;
-using System.Text;
 
 namespace FSH.Modules.Identity.Services;
 
@@ -62,591 +47,8 @@ internal sealed partial class UserService(
     {
         if (string.IsNullOrWhiteSpace(multiTenantContextAccessor?.MultiTenantContext?.TenantInfo?.Id))
         {
-            throw new UnauthorizedException("invalid tenant");
+            throw new FSH.Framework.Core.Exceptions.UnauthorizedException("invalid tenant");
         }
-    }
-
-    public async Task<string> ConfirmEmailAsync(string userId, string code, string tenant, CancellationToken cancellationToken)
-    {
-        EnsureValidTenant();
-
-        var user = await userManager.Users
-            .Where(u => u.Id == userId && !u.EmailConfirmed)
-            .FirstOrDefaultAsync(cancellationToken);
-
-        _ = user ?? throw new CustomException("An error occurred while confirming E-Mail.");
-
-        code = Encoding.UTF8.GetString(WebEncoders.Base64UrlDecode(code));
-        var result = await userManager.ConfirmEmailAsync(user, code);
-
-        return result.Succeeded
-            ? string.Format(CultureInfo.InvariantCulture, "Account Confirmed for E-Mail {0}. You can now use the /api/tokens endpoint to generate JWT.", user.Email)
-            : throw new CustomException(string.Format(CultureInfo.InvariantCulture, "An error occurred while confirming {0}", user.Email));
-    }
-
-    public async Task<string> ConfirmPhoneNumberAsync(string userId, string code)
-    {
-        EnsureValidTenant();
-
-        var user = await userManager.Users
-            .Where(u => u.Id == userId && !u.PhoneNumberConfirmed)
-            .FirstOrDefaultAsync();
-
-        _ = user ?? throw new CustomException("An error occurred while confirming phone number.");
-
-        code = Encoding.UTF8.GetString(WebEncoders.Base64UrlDecode(code));
-        var result = await userManager.ChangePhoneNumberAsync(user, user.PhoneNumber!, code);
-
-        return result.Succeeded
-            ? string.Format(CultureInfo.InvariantCulture, "Phone number {0} confirmed successfully.", user.PhoneNumber)
-            : throw new CustomException(string.Format(CultureInfo.InvariantCulture, "An error occurred while confirming phone number {0}", user.PhoneNumber));
-    }
-
-    public async Task<bool> ExistsWithEmailAsync(string email, string? exceptId = null)
-    {
-        EnsureValidTenant();
-        return await userManager.FindByEmailAsync(email.Normalize()) is FshUser user && user.Id != exceptId;
-    }
-
-    public async Task<bool> ExistsWithNameAsync(string name)
-    {
-        EnsureValidTenant();
-        return await userManager.FindByNameAsync(name) is not null;
-    }
-
-    public async Task<bool> ExistsWithPhoneNumberAsync(string phoneNumber, string? exceptId = null)
-    {
-        EnsureValidTenant();
-        return await userManager.Users.FirstOrDefaultAsync(x => x.PhoneNumber == phoneNumber) is FshUser user && user.Id != exceptId;
-    }
-
-    public async Task<UserDto> GetAsync(string userId, CancellationToken cancellationToken)
-    {
-        var user = await userManager.Users
-            .AsNoTracking()
-            .Where(u => u.Id == userId)
-            .FirstOrDefaultAsync(cancellationToken);
-
-        _ = user ?? throw new NotFoundException("user not found");
-
-        return new UserDto
-        {
-            Id = user.Id,
-            Email = user.Email,
-            UserName = user.UserName,
-            FirstName = user.FirstName,
-            LastName = user.LastName,
-            ImageUrl = ResolveImageUrl(user.ImageUrl),
-            IsActive = user.IsActive
-        };
-    }
-
-    public Task<int> GetCountAsync(CancellationToken cancellationToken) =>
-        userManager.Users.AsNoTracking().CountAsync(cancellationToken);
-
-    public async Task<List<UserDto>> GetListAsync(CancellationToken cancellationToken)
-    {
-        var users = await userManager.Users.AsNoTracking().ToListAsync(cancellationToken);
-        var result = new List<UserDto>(users.Count);
-        foreach (var user in users)
-        {
-            result.Add(new UserDto
-            {
-                Id = user.Id,
-                Email = user.Email,
-                UserName = user.UserName,
-                FirstName = user.FirstName,
-                LastName = user.LastName,
-                ImageUrl = ResolveImageUrl(user.ImageUrl),
-                IsActive = user.IsActive
-            });
-        }
-
-        return result;
-    }
-
-    public async Task<string> GetOrCreateFromPrincipalAsync(ClaimsPrincipal principal)
-    {
-        EnsureValidTenant();
-        ArgumentNullException.ThrowIfNull(principal);
-
-        var email = principal.FindFirstValue(ClaimTypes.Email)
-            ?? principal.FindFirstValue("email")
-            ?? throw new CustomException("Email claim is required for external authentication.");
-
-        // Try to find existing user by email
-        var user = await userManager.FindByEmailAsync(email);
-        if (user is not null)
-        {
-            return user.Id;
-        }
-
-        // Extract claims for new user creation
-        var firstName = principal.FindFirstValue(ClaimTypes.GivenName)
-            ?? principal.FindFirstValue("given_name")
-            ?? string.Empty;
-
-        var lastName = principal.FindFirstValue(ClaimTypes.Surname)
-            ?? principal.FindFirstValue("family_name")
-            ?? string.Empty;
-
-        var userName = principal.FindFirstValue(ClaimTypes.Name)
-            ?? principal.FindFirstValue("preferred_username")
-            ?? email.Split('@')[0];
-
-        // Ensure unique username
-        if (await userManager.FindByNameAsync(userName) is not null)
-        {
-            userName = $"{userName}_{Guid.NewGuid():N}"[..20];
-        }
-
-        // Create new user from external principal
-        user = new FshUser
-        {
-            Email = email,
-            UserName = userName,
-            FirstName = firstName,
-            LastName = lastName,
-            EmailConfirmed = true, // External provider has verified the email
-            PhoneNumberConfirmed = false,
-            IsActive = true
-        };
-
-        var result = await userManager.CreateAsync(user);
-        if (!result.Succeeded)
-        {
-            var errors = result.Errors.Select(e => e.Description).ToList();
-            throw new CustomException("Failed to create user from external principal.", errors);
-        }
-
-        // Assign basic role
-        await userManager.AddToRoleAsync(user, RoleConstants.Basic);
-
-        // Add to default groups
-        var defaultGroups = await db.Groups
-            .Where(g => g.IsDefault && !g.IsDeleted)
-            .ToListAsync();
-
-        foreach (var group in defaultGroups)
-        {
-            db.UserGroups.Add(UserGroup.Create(user.Id, group.Id, "ExternalAuth"));
-        }
-
-        // Raise domain event for user registration
-        var tenantId = multiTenantContextAccessor.MultiTenantContext.TenantInfo?.Id;
-        user.RecordRegistered(tenantId);
-
-        // Save to dispatch domain event via interceptor
-        await db.SaveChangesAsync();
-
-        // Publish integration event
-        var integrationEvent = new UserRegisteredIntegrationEvent(
-            Id: Guid.NewGuid(),
-            OccurredOnUtc: DateTime.UtcNow,
-            TenantId: tenantId,
-            CorrelationId: Guid.NewGuid().ToString(),
-            Source: "Identity.ExternalAuth",
-            UserId: user.Id,
-            Email: user.Email ?? string.Empty,
-            FirstName: user.FirstName ?? string.Empty,
-            LastName: user.LastName ?? string.Empty);
-
-        await outboxStore.AddAsync(integrationEvent).ConfigureAwait(false);
-
-        return user.Id;
-    }
-
-    public async Task<string> RegisterAsync(string firstName, string lastName, string email, string userName, string password, string confirmPassword, string phoneNumber, string origin, CancellationToken cancellationToken)
-    {
-        if (password != confirmPassword) throw new CustomException("password mismatch.");
-
-        // create user entity
-        var user = new FshUser
-        {
-            Email = email,
-            FirstName = firstName,
-            LastName = lastName,
-            UserName = userName,
-            PhoneNumber = phoneNumber,
-            IsActive = true,
-            EmailConfirmed = false,
-            PhoneNumberConfirmed = false,
-        };
-
-        // register user
-        var result = await userManager.CreateAsync(user, password);
-        if (!result.Succeeded)
-        {
-            var errors = result.Errors.Select(error => error.Description).ToList();
-            throw new CustomException("error while registering a new user", errors);
-        }
-
-        // add basic role
-        await userManager.AddToRoleAsync(user, RoleConstants.Basic);
-
-        // add user to default groups
-        var defaultGroups = await db.Groups
-            .Where(g => g.IsDefault && !g.IsDeleted)
-            .ToListAsync(cancellationToken);
-
-        foreach (var group in defaultGroups)
-        {
-            db.UserGroups.Add(UserGroup.Create(user.Id, group.Id, "System"));
-        }
-
-        if (defaultGroups.Count > 0)
-        {
-            await db.SaveChangesAsync(cancellationToken);
-        }
-
-        // send confirmation mail
-        if (!string.IsNullOrEmpty(user.Email))
-        {
-            string emailVerificationUri = await GetEmailVerificationUriAsync(user, origin);
-            string emailBody = BuildConfirmationEmailHtml(user.FirstName ?? user.UserName ?? "User", emailVerificationUri);
-            var mailRequest = new MailRequest(
-                new Collection<string> { user.Email },
-                "Confirm Your Email Address",
-                emailBody);
-            jobService.Enqueue("email", () => mailService.SendAsync(mailRequest, cancellationToken));
-        }
-
-        // Raise domain event for user registration
-        var tenantId = multiTenantContextAccessor.MultiTenantContext.TenantInfo?.Id;
-        user.RecordRegistered(tenantId);
-
-        // Save to dispatch domain event via interceptor
-        await db.SaveChangesAsync(cancellationToken);
-
-        // enqueue integration event for user registration
-        var correlationId = Guid.NewGuid().ToString();
-        var integrationEvent = new UserRegisteredIntegrationEvent(
-            Id: Guid.NewGuid(),
-            OccurredOnUtc: DateTime.UtcNow,
-            TenantId: tenantId,
-            CorrelationId: correlationId,
-            Source: "Identity",
-            UserId: user.Id,
-            Email: user.Email ?? string.Empty,
-            FirstName: user.FirstName ?? string.Empty,
-            LastName: user.LastName ?? string.Empty);
-
-        await outboxStore.AddAsync(integrationEvent, cancellationToken).ConfigureAwait(false);
-
-        return user.Id;
-    }
-
-    public async Task ToggleStatusAsync(bool activateUser, string userId, CancellationToken cancellationToken)
-    {
-        EnsureValidTenant();
-
-        var actorId = _currentUser.GetUserId();
-        if (actorId == Guid.Empty)
-        {
-            throw new UnauthorizedException("authenticated user required to toggle status");
-        }
-
-        var actor = await userManager.FindByIdAsync(actorId.ToString());
-        _ = actor ?? throw new UnauthorizedException("current user not found");
-
-        async ValueTask AuditPolicyFailureAsync(string reason, CancellationToken ct)
-        {
-            var tenant = multiTenantContextAccessor?.MultiTenantContext?.TenantInfo?.Id ?? "unknown";
-            var claims = new Dictionary<string, object?>
-            {
-                ["actorId"] = actorId.ToString(),
-                ["targetUserId"] = userId,
-                ["tenant"] = tenant,
-                ["action"] = activateUser ? "activate" : "deactivate"
-            };
-
-            await _auditClient.WriteSecurityAsync(
-                SecurityAction.PolicyFailed,
-                subjectId: actorId.ToString(),
-                reasonCode: reason,
-                claims: claims,
-                severity: AuditSeverity.Warning,
-                source: "Identity",
-                ct: ct).ConfigureAwait(false);
-        }
-
-        if (!await userManager.IsInRoleAsync(actor, RoleConstants.Admin))
-        {
-            await AuditPolicyFailureAsync("ActorNotAdmin", cancellationToken);
-            throw new CustomException("Only administrators can toggle user status.");
-        }
-
-        if (!activateUser && string.Equals(actor.Id, userId, StringComparison.Ordinal))
-        {
-            await AuditPolicyFailureAsync("SelfDeactivationBlocked", cancellationToken);
-            throw new CustomException("Users cannot deactivate themselves.");
-        }
-
-        var user = await userManager.Users.Where(u => u.Id == userId).FirstOrDefaultAsync(cancellationToken);
-        _ = user ?? throw new NotFoundException("User Not Found.");
-
-        bool targetIsAdmin = await userManager.IsInRoleAsync(user, RoleConstants.Admin);
-        if (targetIsAdmin)
-        {
-            await AuditPolicyFailureAsync("AdminDeactivationBlocked", cancellationToken);
-            throw new CustomException("Administrators cannot be deactivated.");
-        }
-
-        if (!activateUser)
-        {
-            var activeAdmins = await userManager.GetUsersInRoleAsync(RoleConstants.Admin);
-            int activeAdminCount = activeAdmins.Count(u => u.IsActive);
-            if (activeAdminCount == 0)
-            {
-                await AuditPolicyFailureAsync("NoActiveAdmins", cancellationToken);
-                throw new CustomException("Tenant must have at least one active administrator.");
-            }
-        }
-
-        var tenantId = multiTenantContextAccessor?.MultiTenantContext?.TenantInfo?.Id;
-        if (activateUser)
-        {
-            user.Activate(actorId.ToString(), tenantId);
-        }
-        else
-        {
-            user.Deactivate(actorId.ToString(), "Status toggled by administrator", tenantId);
-        }
-
-        var result = await userManager.UpdateAsync(user);
-        if (!result.Succeeded)
-        {
-            var errors = result.Errors.Select(error => error.Description).ToList();
-            throw new CustomException("Toggle status failed", errors);
-        }
-
-        await _auditClient.WriteActivityAsync(
-            ActivityKind.Command,
-            name: "ToggleUserStatus",
-            statusCode: 204,
-            durationMs: 0,
-            captured: BodyCapture.None,
-            requestSize: 0,
-            responseSize: 0,
-            requestPreview: new { actorId = actorId.ToString(), targetUserId = userId, action = activateUser ? "activate" : "deactivate", tenant = tenantId ?? "unknown" },
-            responsePreview: new { outcome = "success" },
-            severity: AuditSeverity.Information,
-            source: "Identity",
-            ct: cancellationToken).ConfigureAwait(false);
-    }
-
-    public async Task UpdateAsync(string userId, string firstName, string lastName, string phoneNumber, FileUploadRequest image, bool deleteCurrentImage)
-    {
-        var user = await userManager.FindByIdAsync(userId);
-
-        _ = user ?? throw new NotFoundException("user not found");
-
-        Uri imageUri = user.ImageUrl ?? null!;
-        if (image.Data != null || deleteCurrentImage)
-        {
-            var imageString = await storageService.UploadAsync<FshUser>(image, FileType.Image);
-            user.ImageUrl = new Uri(imageString, UriKind.RelativeOrAbsolute);
-            if (deleteCurrentImage && imageUri != null)
-            {
-                await storageService.RemoveAsync(imageUri.ToString());
-            }
-        }
-
-        user.FirstName = firstName;
-        user.LastName = lastName;
-        string? currentPhoneNumber = await userManager.GetPhoneNumberAsync(user);
-        if (phoneNumber != currentPhoneNumber)
-        {
-            await userManager.SetPhoneNumberAsync(user, phoneNumber);
-        }
-
-        var result = await userManager.UpdateAsync(user);
-        await signInManager.RefreshSignInAsync(user);
-
-        if (!result.Succeeded)
-        {
-            throw new CustomException("Update profile failed");
-        }
-    }
-
-    public async Task DeleteAsync(string userId)
-    {
-        FshUser? user = await userManager.FindByIdAsync(userId);
-
-        _ = user ?? throw new NotFoundException("User Not Found.");
-
-        user.IsActive = false;
-        IdentityResult? result = await userManager.UpdateAsync(user);
-
-        if (!result.Succeeded)
-        {
-            List<string> errors = result.Errors.Select(error => error.Description).ToList();
-            throw new CustomException("Delete profile failed", errors);
-        }
-    }
-
-    private async Task<string> GetEmailVerificationUriAsync(FshUser user, string origin)
-    {
-        EnsureValidTenant();
-
-        string code = await userManager.GenerateEmailConfirmationTokenAsync(user);
-        code = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
-        const string route = "api/v1/identity/confirm-email";
-        var endpointUri = new Uri(string.Concat($"{origin}/", route));
-        string verificationUri = QueryHelpers.AddQueryString(endpointUri.ToString(), QueryStringKeys.UserId, user.Id);
-        verificationUri = QueryHelpers.AddQueryString(verificationUri, QueryStringKeys.Code, code);
-        verificationUri = QueryHelpers.AddQueryString(verificationUri,
-            MultitenancyConstants.Identifier,
-            multiTenantContextAccessor?.MultiTenantContext?.TenantInfo?.Id!);
-        return verificationUri;
-    }
-
-    private static string BuildConfirmationEmailHtml(string userName, string confirmationUrl)
-    {
-        return $"""
-            <!DOCTYPE html>
-            <html>
-            <head>
-                <meta charset="utf-8">
-                <meta name="viewport" content="width=device-width, initial-scale=1.0">
-                <title>Confirm Your Email</title>
-            </head>
-            <body style="margin: 0; padding: 0; font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif; background-color: #f8fafc;">
-                <table role="presentation" style="width: 100%; border-collapse: collapse;">
-                    <tr>
-                        <td align="center" style="padding: 40px 0;">
-                            <table role="presentation" style="width: 100%; max-width: 600px; border-collapse: collapse; background-color: #ffffff; border-radius: 8px; box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);">
-                                <tr>
-                                    <td style="padding: 40px 40px 30px 40px; text-align: center; background-color: #2563eb; border-radius: 8px 8px 0 0;">
-                                        <h1 style="margin: 0; color: #ffffff; font-size: 24px; font-weight: 600;">Confirm Your Email Address</h1>
-                                    </td>
-                                </tr>
-                                <tr>
-                                    <td style="padding: 40px;">
-                                        <p style="margin: 0 0 20px 0; color: #334155; font-size: 16px; line-height: 1.6;">
-                                            Hi {System.Net.WebUtility.HtmlEncode(userName)},
-                                        </p>
-                                        <p style="margin: 0 0 20px 0; color: #334155; font-size: 16px; line-height: 1.6;">
-                                            Thank you for registering! Please confirm your email address by clicking the button below:
-                                        </p>
-                                        <table role="presentation" style="width: 100%; border-collapse: collapse;">
-                                            <tr>
-                                                <td align="center" style="padding: 30px 0;">
-                                                    <a href="{System.Net.WebUtility.HtmlEncode(confirmationUrl)}" style="display: inline-block; padding: 14px 32px; background-color: #2563eb; color: #ffffff; text-decoration: none; font-size: 16px; font-weight: 600; border-radius: 6px;">
-                                                        Confirm Email Address
-                                                    </a>
-                                                </td>
-                                            </tr>
-                                        </table>
-                                        <p style="margin: 0 0 20px 0; color: #64748b; font-size: 14px; line-height: 1.6;">
-                                            If the button doesn't work, copy and paste this link into your browser:
-                                        </p>
-                                        <p style="margin: 0 0 20px 0; color: #2563eb; font-size: 14px; line-height: 1.6; word-break: break-all;">
-                                            {System.Net.WebUtility.HtmlEncode(confirmationUrl)}
-                                        </p>
-                                        <p style="margin: 30px 0 0 0; color: #64748b; font-size: 14px; line-height: 1.6;">
-                                            If you didn't create an account, you can safely ignore this email.
-                                        </p>
-                                    </td>
-                                </tr>
-                                <tr>
-                                    <td style="padding: 20px 40px; background-color: #f1f5f9; border-radius: 0 0 8px 8px; text-align: center;">
-                                        <p style="margin: 0; color: #94a3b8; font-size: 12px;">
-                                            This is an automated message. Please do not reply to this email.
-                                        </p>
-                                    </td>
-                                </tr>
-                            </table>
-                        </td>
-                    </tr>
-                </table>
-            </body>
-            </html>
-            """;
-    }
-
-    public async Task<string> AssignRolesAsync(string userId, List<UserRoleDto> userRoles, CancellationToken cancellationToken)
-    {
-        var user = await userManager.Users.Where(u => u.Id == userId).FirstOrDefaultAsync(cancellationToken);
-
-        _ = user ?? throw new NotFoundException("user not found");
-
-        // Check if the user is an admin for which the admin role is getting disabled
-        if (await userManager.IsInRoleAsync(user, RoleConstants.Admin)
-            && userRoles.Exists(a => !a.Enabled && a.RoleName == RoleConstants.Admin))
-        {
-            // Get count of users in Admin Role
-            int adminCount = (await userManager.GetUsersInRoleAsync(RoleConstants.Admin)).Count;
-
-            // Check if user is not Root Tenant Admin
-            // Edge Case : there are chances for other tenants to have users with the same email as that of Root Tenant Admin. Probably can add a check while User Registration
-            if (user.Email == MultitenancyConstants.Root.EmailAddress)
-            {
-                if (multiTenantContextAccessor?.MultiTenantContext?.TenantInfo?.Id == MultitenancyConstants.Root.Id)
-                {
-                    throw new CustomException("action not permitted");
-                }
-            }
-            else if (adminCount <= 2)
-            {
-                throw new CustomException("tenant should have at least 2 admins.");
-            }
-        }
-
-        var assignedRoles = new List<string>();
-
-        foreach (var userRole in userRoles)
-        {
-            // Check if Role Exists
-            if (await roleManager.FindByNameAsync(userRole.RoleName!) is not null)
-            {
-                if (userRole.Enabled)
-                {
-                    if (!await userManager.IsInRoleAsync(user, userRole.RoleName!))
-                    {
-                        await userManager.AddToRoleAsync(user, userRole.RoleName!);
-                        assignedRoles.Add(userRole.RoleName!);
-                    }
-                }
-                else
-                {
-                    await userManager.RemoveFromRoleAsync(user, userRole.RoleName!);
-                }
-            }
-        }
-
-        // Raise domain event for newly assigned roles
-        if (assignedRoles.Count > 0)
-        {
-            var tenantId = multiTenantContextAccessor?.MultiTenantContext?.TenantInfo?.Id;
-            user.RecordRolesAssigned(assignedRoles, tenantId);
-            await db.SaveChangesAsync(cancellationToken);
-        }
-
-        return "User Roles Updated Successfully.";
-
-    }
-
-    public async Task<List<UserRoleDto>> GetUserRolesAsync(string userId, CancellationToken cancellationToken)
-    {
-        var userRoles = new List<UserRoleDto>();
-
-        var user = await userManager.FindByIdAsync(userId);
-        if (user is null) throw new NotFoundException("user not found");
-        var roles = await roleManager.Roles.AsNoTracking().ToListAsync(cancellationToken);
-        if (roles is null) throw new NotFoundException("roles not found");
-        foreach (var role in roles)
-        {
-            userRoles.Add(new UserRoleDto
-            {
-                RoleId = role.Id,
-                RoleName = role.Name,
-                Description = role.Description,
-                Enabled = await userManager.IsInRoleAsync(user, role.Name!)
-            });
-        }
-
-        return userRoles;
     }
 
     private string? ResolveImageUrl(Uri? imageUrl)
